Иду сдавать 3-ю лабу

+++ labs/lab3/misha/src/chart.cpp
@@ -28,25 +28,25 @@ void Chart::clear() {
     callDataChanged();
 }
 
-std::vector<double> Chart::calcCommonAccelerations() {
-    std::vector<double> result = calcGradV();
-    for (double &val: result) {
+std::vector<long double> Chart::calcCommonAccelerations() {
+    std::vector<long double> result = calcGradV();
+    for (long double &val: result) {
         val *= -1.0 / mass;
     }
     return result;
 }
 
-double Chart::calcHamiltonian() {
-    double result = 0;
-    std::vector<double> V = calcV();
+long double Chart::calcHamiltonian() {
+    long double result = 0;
+    std::vector<long double> V = calcV();
     for (size_t i = 0; i < V.size(); ++i) {
         result += speeds[i] * speeds[i] * mass / 2 + V[i];
     }
     return result;
 }
 
-std::vector<double> Chart::calcGradV() {
-    std::vector<double> result(numberOfParticles);
+std::vector<long double> Chart::calcGradV() {
+    std::vector<long double> result(numberOfParticles);
 
     result[0] = calcGradV(offsets[numberOfParticles - 1], offsets[0], offsets[1]);
     result[numberOfParticles - 1] = calcGradV(offsets[numberOfParticles - 2],
@@ -60,14 +60,14 @@ std::vector<double> Chart::calcGradV() {
     return result;
 }
 
-double Chart::calcGradV(double qM1, double q, double qP1) {
+long double Chart::calcGradV(long double qM1,long  double q,long  double qP1) {
     return -(qP1 - q) + (q - qM1) +
            alpha * (-std::pow(qP1 - q, 2) + std::pow(q - qM1, 2)) +
            beta * (-std::pow(qP1 - q, 3) + std::pow(q - qM1, 3));
 }
 
-std::vector<double> Chart::calcV() {
-    std::vector<double> result(offsets.size());
+std::vector<long double> Chart::calcV() {
+    std::vector<long double> result(offsets.size());
     for (size_t i = 0; i < result.size() - 1; ++i) {
         result[i] = calcV(offsets[i], offsets[i + 1]);
     }
@@ -75,7 +75,7 @@ std::vector<double> Chart::calcV() {
     return result;
 }
 
-double Chart::calcV(double q, double qP1) {
+long double Chart::calcV(long double q, long double qP1) {
     double r = qP1 - q;
     return r * r / 2.0 + alpha * r * r * r / 3.0 + beta * r * r * r * r / 4.0;
 }
@@ -91,7 +91,7 @@ void Chart::callDataChanged() {
 }
 
 
-void Chart::setAlphaBeta(double alpha, double beta) {
+void Chart::setAlphaBeta(long double alpha,long  double beta) {
     this->alpha = alpha;
     this->beta = beta;
 }
